Move test file helpers
