+++ pkg/acceptance/helpers/tmp_toml_setup_helpers.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/helpers/random"
-	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/testhelpers"
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/testfiles"
 )
 
 // TODO [SNOW-1827324]: add TestClient ref to each specific client, so that we enhance specific client and not the base one
@@ -106,7 +106,7 @@ func (c *TestClient) StoreTempTomlConfigWithProfile(t *testing.T, profile string
 	t.Helper()
 
 	toml := tomlProvider(profile)
-	configPath := testhelpers.TestFile(t, random.AlphaN(10), []byte(toml))
+	configPath := testfiles.TestFile(t, random.AlphaN(10), []byte(toml))
 	return &TmpTomlConfig{
 		Profile: profile,
 		Path:    configPath,
@@ -118,7 +118,7 @@ func (c *TestClient) StoreTempTomlConfigWithCustomPermissions(t *testing.T, toml
 
 	profile := random.AlphaN(6)
 	toml := tomlProvider(profile)
-	configPath := testhelpers.TestFileWithCustomPermissions(t, random.AlphaN(10), []byte(toml), permissions)
+	configPath := testfiles.TestFileWithCustomPermissions(t, random.AlphaN(10), []byte(toml), permissions)
 	return &TmpTomlConfig{
 		Profile: profile,
 		Path:    configPath,
