+++ pkg/acceptance/testfiles/helpers.go
@@ -1,4 +1,4 @@
-package testhelpers
+package testfiles
 
 import (
 	"io/fs"
@@ -10,7 +10,6 @@ import (
 	"golang.org/x/sys/unix"
 )
 
-// TODO [next PRs]: move this file
 // TestFile creates a temporary file with the given filename, data and with the default permissions.
 // The directory is automatically removed when the test and all its subtests complete.
 // Each subsequent call to t.TempDir returns a unique directory.
