+++ pkg/manual_tests/windows/files_test.go
@@ -3,8 +3,8 @@ package windows_test
 import (
 	"testing"
 
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/testfiles"
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/internal/oswrapper"
-	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/testhelpers"
 	"github.com/stretchr/testify/require"
 )
 
@@ -13,7 +13,7 @@ func TestReadFileSafeWorksOnWindows(t *testing.T) {
 		t.Skip("reading files on other platforms is currently done in the sdk package")
 	}
 	exp := []byte("content")
-	configPath := testhelpers.TestFile(t, "config", exp)
+	configPath := testfiles.TestFile(t, "config", exp)
 
 	act, err := oswrapper.ReadFileSafe(configPath, true)
 	require.NoError(t, err)
